Add Makefile (WIP)
This adds a Makefile to be able to build the project independently.

But this Makefile is made in a way that if it's called from within Poke_Transporter_GB, it works as well.

But right now we still have a conflict regarding the pokemon_data.

I will have to rework this a bit in order to unify the workflow for PTGB and this project as an independent lib.