Always generate tables as .bin files and inject them into the lib.
Split off table-generator from data-generator in Poke_Transporter_GB. It will only be used to generate the tables now.
We don't want the other dependencies from Poke_Transporter_GB here. This is the only stuff that's relevant here.

We always inject them into the lib. We no longer maintain a separate copy of the same data in the source code.

The only difference between running on gba and elsewhere, is that the tables will only get compressed for GBA.