Fix compile error caused by dependency issue on bin2s_tool when run with -j
When building with multiple parallel jobs, you could end up in a situation in which bin2s_tool is getting built
in parallel alongside the source files of the lib before the .bin files have been transformed into object files.

Fix: make sure bin2s is built in the first stage of the Makefile (alongside generate_data) before the rest of the lib gets built.