Fix compatibility with MSYS2 MinGW64 on Windows
The biggest problem was that the TMPDIR var did not propagate to the table-generator sub-make,
because we nuked the environment with env -i because table-generator needs to be built with the
host compiler, not the target compiler. This works fine on Linux, WSL2 or even the
regular MSYS2 terminal.

But it fails with the MSYS2 MinGW64 terminal.

The solution is to use env - and pass the relevant variables to it to ensure the temp folder is
properly set.